Fix signature issue with AbstractContainer::offsetGet
`AbstractContainer` inherits from `Zend\Stdlib\ArrayObject`.
`ArrayObject::offsetGet()` is defined such as to _return by reference_;
however, the override in `AbstractContainer` did not. This casuses an
issue in current alpha versions of PHP 7.2, as it is considered an
incompatible signature.

This patch updates the `AbstractContainer::offsetGet()` declaration to
return by reference. All tests related to containers continue to pass.