Mock PHP built-ins in order to test setting known save handler
Adds php-mock/php-mock-phpunit as a dev dependency, and uses it within
the `SessionConfig` unit tests to mock both `phpinfo()` and
`session_module_name()` in order to validate that the method can be used
to set arbitrary session save handlers so long as they are supported by
the engine.

PHP 5.6 requires a lower version of php-mock-phpunit, so the package is
added to the `LEGACY_DEPS` environment for the 5.6 locked job, and we
allow either 1.1.2 or 2.0 versions of the package within its
constraints to allow the other PHP 5.6 versions to receive that version.