Properly merge arrays of allowed hash algorithms.
The `+` operator does not renumber array keys.
The keys that existed on the left side will replace fields from the
right hand side with the same keys.

On my installation, `md2` and `md4` would left out of the final result.

This was detected by static analysis. I don't have a personal use case
for md2/md4.