Fixed sessions not being destroyed on expire by db table gateway
When reading an expired session in the db table gateway they should
be removed right away by the destroy method. This doesn't really
happen because the destroy method uses the read method to see if
the session exists only to conclude the session doens't exist because
the read method returns false for expired sessions even when they
are still in the database.

Because a delete is an atomic operation that does not throw errors
when there is nothing to delete we should actually always just run
the delete operation.