qa: add `ReturnTypeWillChange` stub
With PHP 8.1, an attribute `ReturnTypeWillChange` was added.
The coding standard does not detect that attribute as a class and uses `const` as import statement for it. To avoid this, I've added the proper usage and annotated it in the class PHPDoc so the coding standard can properly detect its usage as otherwise it would be removed.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>