qa: skip tests that were validating deprecations
PHPUnit 9.5.10+ makes these tests fail, even when `convertDeprecationsToExceptions` is enabled; we need to skip them to make the test suite pass.
Since the tests were halting execution of these methods on previous PHPUnit versions as soon as the deprecation was triggered, they were essentially no-ops, and can be safely skipped.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>